Add #includes for OpenBSD and NetBSD compat.
Danglish-out-stompification from github pull requests.

+++ param.c
@@ -55,7 +55,7 @@ param_wrapline(struct ocx *ocx, const char *b)
 	Put(ocx, OCX_DIAG, "\t");
 	e = strchr(b, '\0');
 	while (b < e) {
-		if (!isspace(*b)) {
+		if (!isspace((int)*b)) {
 			Put(ocx, OCX_DIAG, "%c", *b);
 			b++;
 			n++;
@@ -72,7 +72,7 @@ param_wrapline(struct ocx *ocx, const char *b)
 		} else {
 			assert (*b == ' ');
 			for (w = b + 1; w < e; w++)
-				if (isspace(*w))
+				if (isspace((int)*w))
 					break;
 			if (n + (w - b) < wrap_at) {
 				Put(ocx, OCX_DIAG, "%.*s", (int)(w - b), b);
